requirements.txt
python-telegram-bot>=13.0
yt-dlp
